extract class Game to separate file
